refactor: common 패키지 global 패키지로 통합 #131
