--- src/main/java/com/bookripple/api/common/code/BaseErrorCode.java
@@ -1,11 +1,0 @@
-package com.bookripple.api.common.code;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseErrorCode {
-    HttpStatus getHttpStatus();
-
-    String getCode();
-
-    String getMessage();
-}
