--- src/main/java/com/bookripple/api/common/code/BaseSuccessCode.java
@@ -1,11 +1,0 @@
-package com.bookripple.api.common.code;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseSuccessCode {
-    HttpStatus getHttpStatus();
-
-    String getCode();
-
-    String getMessage();
-}
