+++ src/main/java/com/bookripple/api/global/code/BaseErrorCode.java
@@ -1,0 +1,12 @@
+package com.bookripple.api.global.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseErrorCode {
+
+  HttpStatus getHttpStatus();
+
+  String getCode();
+
+  String getMessage();
+}
