+++ src/main/java/com/bookripple/api/global/response/ApiResponse.java
@@ -1,0 +1,34 @@
+package com.bookripple.api.global.response;
+
+import com.bookripple.api.global.code.BaseErrorCode;
+import com.bookripple.api.global.code.BaseSuccessCode;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+@JsonPropertyOrder({"isSuccess", "code", "message", "result"})
+public class ApiResponse<T> {
+
+  @JsonProperty("isSuccess")
+  private final Boolean isSuccess;
+
+  @JsonProperty("code")
+  private final String code;
+
+  @JsonProperty("message")
+  private final String message;
+
+  @JsonProperty("result")
+  private T result;
+
+  public static <T> ApiResponse<T> onSuccess(BaseSuccessCode code, T result) {
+    return new ApiResponse<>(true, code.getCode(), code.getMessage(), result);
+  }
+
+  public static <T> ApiResponse<T> onFailure(BaseErrorCode code, T result) {
+    return new ApiResponse<>(false, code.getCode(), code.getMessage(), result);
+  }
+}
